backend/requirements.txt
# AI Tutor Backend Dependencies
fastapi>=0.115
uvicorn[standard]>=0.30
sse-starlette>=2.0
python-multipart>=0.0.9

# LLM & Embeddings
openai>=1.40

# Database
sqlalchemy>=2.0
psycopg[binary]>=3.3
pgvector>=0.3

# Document extraction
pdfplumber>=0.11
python-docx>=1.1
PyMuPDF>=1.24

# Config
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0

# Auth (SuperTokens)
supertokens-python>=0.27
